Move command line interface utilities out to dedicated files
