+++ website/src/main/kotlin/org/pointyware/xyz/site/utils/ProgramInputs.kt
@@ -1,0 +1,10 @@
+package org.pointyware.xyz.site.utils
+
+/**
+ * Defines the available parameters for controlling program behavior.
+ * This model exists primarily as a helper object for processing arguments progressively.
+ * @see CommandOption
+ */
+data class ProgramInputs(
+    val output: ProgramOutput
+)
